Module aws_sdk_ioteventsdata::model

Modules

acknowledge_action_configuration

See AcknowledgeActionConfiguration

acknowledge_alarm_action_request

See AcknowledgeAlarmActionRequest

alarm

See Alarm

alarm_state

See AlarmState

alarm_summary

See AlarmSummary

batch_alarm_action_error_entry

See BatchAlarmActionErrorEntry

batch_put_message_error_entry

See BatchPutMessageErrorEntry

batch_update_detector_error_entry

See BatchUpdateDetectorErrorEntry

customer_action

See CustomerAction

detector

See Detector

detector_state

See DetectorState

detector_state_definition

See DetectorStateDefinition

detector_state_summary

See DetectorStateSummary

detector_summary

See DetectorSummary

disable_action_configuration

See DisableActionConfiguration

disable_alarm_action_request

See DisableAlarmActionRequest

enable_action_configuration

See EnableActionConfiguration

enable_alarm_action_request

See EnableAlarmActionRequest

message

See Message

reset_action_configuration

See ResetActionConfiguration

reset_alarm_action_request

See ResetAlarmActionRequest

rule_evaluation

See RuleEvaluation

simple_rule_evaluation

See SimpleRuleEvaluation

snooze_action_configuration

See SnoozeActionConfiguration

snooze_alarm_action_request

See SnoozeAlarmActionRequest

state_change_configuration

See StateChangeConfiguration

system_event

See SystemEvent

timer

See Timer

timer_definition

See TimerDefinition

timestamp_value

See TimestampValue

update_detector_request

See UpdateDetectorRequest

variable

See Variable

variable_definition

See VariableDefinition

Structs

AcknowledgeActionConfiguration

Contains the configuration information of an acknowledge action.

AcknowledgeAlarmActionRequest

Information needed to acknowledge the alarm.

Alarm

Contains information about an alarm.

AlarmState

Contains information about the current state of the alarm.

AlarmSummary

Contains a summary of an alarm.

BatchAlarmActionErrorEntry

Contains error messages associated with one of the following requests:

BatchPutMessageErrorEntry

Contains information about the errors encountered.

BatchUpdateDetectorErrorEntry

Information about the error that occurred when attempting to update a detector.

CustomerAction

Contains information about the action that you can take to respond to the alarm.

Detector

Information about the detector (instance).

DetectorState

Information about the current state of the detector instance.

DetectorStateDefinition

The new state, variable values, and timer settings of the detector (instance).

DetectorStateSummary

Information about the detector state.

DetectorSummary

Information about the detector (instance).

DisableActionConfiguration

Contains the configuration information of a disable action.

DisableAlarmActionRequest

Information used to disable the alarm.

EnableActionConfiguration

Contains the configuration information of an enable action.

EnableAlarmActionRequest

Information needed to enable the alarm.

Message

Information about a message.

ResetActionConfiguration

Contains the configuration information of a reset action.

ResetAlarmActionRequest

Information needed to reset the alarm.

RuleEvaluation

Information needed to evaluate data.

SimpleRuleEvaluation

Information needed to compare two values with a comparison operator.

SnoozeActionConfiguration

Contains the configuration information of a snooze action.

SnoozeAlarmActionRequest

Information needed to snooze the alarm.

StateChangeConfiguration

Contains the configuration information of alarm state changes.

SystemEvent

Contains information about alarm state changes.

Timer

The current state of a timer.

TimerDefinition

The new setting of a timer.

TimestampValue

Contains information about a timestamp.

UpdateDetectorRequest

Information used to update the detector (instance).

Variable

The current state of the variable.

VariableDefinition

The new value of the variable.

Enums

AlarmStateName

ComparisonOperator

CustomerActionName

ErrorCode

EventType

TriggerType