Module aws_sdk_ioteventsdata::operation

Structs

BatchAcknowledgeAlarm

Acknowledges one or more alarms. The alarms change to the ACKNOWLEDGED state

BatchDisableAlarm

Disables one or more alarms. The alarms change to the DISABLED state after

BatchEnableAlarm

Enables one or more alarms. The alarms change to the NORMAL state after you

BatchPutMessage

Sends a set of messages to the AWS IoT Events system. Each message payload is transformed into

BatchResetAlarm

Resets one or more alarms. The alarms return to the NORMAL state after you

BatchSnoozeAlarm

Changes one or more alarms to the snooze mode. The alarms change to the

BatchUpdateDetector

Updates the state, variable values, and timer settings of one or more detectors

DescribeAlarm

Retrieves information about an alarm.

DescribeDetector

Returns information about the specified detector (instance).

ListAlarms

Lists one or more alarms. The operation returns only the metadata associated with each

ListDetectors

Lists detectors (the instances of a detector model).